fix(log): use size_t for severity index
Change the type of the severity index from uint32_t to size_t in cLog::Write.
This ensures compatibility with the array indexing and avoids potential
warnings or errors on platforms where size_t and uint32_t differ.

+++ src/Log.cpp
@@ -19,6 +19,6 @@ void cLog::Write(Severity severity, const std::string& msg)
         "(DD) ",
     };
 
-    auto idx = static_cast<uint32_t>(severity);
+    auto idx = static_cast<size_t>(severity);
     ::printf("%s%s\n", SeverityNames[idx], msg.c_str());
 }
